// src/components/pages/Interface/types.ts
export interface InterfaceType {
  id: string;
  vpcId: string;
  vpcName: string;
  projectId: string;
  macAddress: string;
  status: string;
  adminState: string;
  ipAddress: string;
  subnetId: string;
  deviceOwner: string;
  deviceId: string;
  vnicType: string;
  createdAt: string;
  updatedAt: string;
}